further file organization, working android build
