+++ lib/descart.dart
@@ -1,0 +1,16 @@
+import 'package:descart/nav_scaffold.dart';
+import 'package:flutter/material.dart';
+
+class DesCart extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Des-Cart',
+      theme: ThemeData(
+        primarySwatch: Colors.green,
+        visualDensity: VisualDensity.adaptivePlatformDensity,
+      ),
+      home: NavScaffold(),
+    );
+  }
+}
