+++ test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:descart/main.dart';
+import 'package:descart/descart.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
